+++ client/vscode/src/backend/instanceVersion.ts
@@ -1,6 +1,7 @@
 import { gql } from '@sourcegraph/http-client'
 import { EventSource } from '@sourcegraph/shared/src/graphql-operations'
 
+import { displayWarning } from '../settings/displayWarnings'
 import { INSTANCE_VERSION_NUMBER_KEY, LocalStorageService } from '../settings/LocalStorageService'
 
 import { requestGraphQLFromVSCode } from './requestGraphQl'
@@ -20,11 +21,17 @@ export function initializeInstantVersionNumber(localStorageService: LocalStorage
                     siteVersionResult.data.site.productVersion.length > 8
                         ? '999999'
                         : siteVersionResult.data.site.productVersion.split('.').join('')
+                if (flattenVersion < '3320') {
+                    displayWarning(
+                        'Your Sourcegraph instance version is not fully compatible with the Sourcegraph extension. Please ask your site admin to upgrade to version 3.32.0 or above. Read more about version support in our [troubleshooting docs](https://docs.sourcegraph.com/admin/how-to/troubleshoot-sg-extension#unsupported-features-by-sourcegraph-version).'
+                    ).catch(() => {})
+                }
                 await localStorageService.setValue(INSTANCE_VERSION_NUMBER_KEY, flattenVersion)
             }
         })
         .catch(error => {
             console.error('Failed to get instance version from host:', error)
+            displayWarning('Cannot determine instance version number').catch(() => {})
         })
     const versionNumber = localStorageService.getValue(INSTANCE_VERSION_NUMBER_KEY)
     // instances below 3.38.0 does not support EventSource.IDEEXTENSION and should fallback to BACKEND source
@@ -32,7 +39,7 @@ export function initializeInstantVersionNumber(localStorageService: LocalStorage
 }
 
 const siteVersionQuery = gql`
-    query {
+    query SiteProductVersion {
         site {
             productVersion
         }